fix(test): move cf func into own file
